+++ daml-script/runner/src/main/scala/com/digitalasset/daml/lf/engine/script/ledgerinteraction/ScriptLedgerClient.scala
@@ -1,0 +1,133 @@
+// Copyright (c) 2021 Digital Asset (Switzerland) GmbH and/or its affiliates. All rights reserved.
+// SPDX-License-Identifier: Apache-2.0
+
+package com.daml.lf.engine.script.ledgerinteraction
+
+import akka.stream.Materializer
+import com.daml.grpc.adapter.ExecutionSequencerFactory
+import com.daml.ledger.api.domain.PartyDetails
+import com.daml.lf.command
+import com.daml.lf.data.Ref._
+import com.daml.lf.data.{Ref, Time}
+import com.daml.lf.speedy.SValue
+import com.daml.lf.value.Value
+import com.daml.lf.value.Value.ContractId
+import io.grpc.StatusRuntimeException
+import scalaz.OneAnd
+
+import scala.concurrent.{ExecutionContext, Future}
+
+// We have our own type for time modes since TimeProviderType
+// allows for more stuff that doesn’t make sense in DAML Script.
+sealed trait ScriptTimeMode
+
+object ScriptTimeMode {
+  final case object Static extends ScriptTimeMode
+  final case object WallClock extends ScriptTimeMode
+}
+
+object ScriptLedgerClient {
+
+  sealed trait CommandResult
+  final case class CreateResult(contractId: ContractId) extends CommandResult
+  final case class ExerciseResult(
+      templateId: Identifier,
+      choice: ChoiceName,
+      result: Value[ContractId],
+  ) extends CommandResult
+
+  final case class ActiveContract(
+      templateId: Identifier,
+      contractId: ContractId,
+      argument: Value[ContractId],
+  )
+
+  final case class TransactionTree(rootEvents: List[TreeEvent])
+  sealed trait TreeEvent
+  final case class Exercised(
+      templateId: Identifier,
+      contractId: ContractId,
+      choice: ChoiceName,
+      argument: Value[ContractId],
+      childEvents: List[TreeEvent],
+  ) extends TreeEvent
+  final case class Created(
+      templateId: Identifier,
+      contractId: ContractId,
+      argument: Value[ContractId],
+  ) extends TreeEvent
+}
+
+// This abstracts over the interaction with the ledger. This allows
+// us to plug in something that interacts with the JSON API as well as
+// something that works against the gRPC API.
+trait ScriptLedgerClient {
+  def query(parties: OneAnd[Set, Ref.Party], templateId: Identifier)(implicit
+      ec: ExecutionContext,
+      mat: Materializer,
+  ): Future[Seq[ScriptLedgerClient.ActiveContract]]
+
+  def queryContractId(parties: OneAnd[Set, Ref.Party], templateId: Identifier, cid: ContractId)(
+      implicit
+      ec: ExecutionContext,
+      mat: Materializer,
+  ): Future[Option[ScriptLedgerClient.ActiveContract]]
+
+  def queryContractKey(
+      parties: OneAnd[Set, Ref.Party],
+      templateId: Identifier,
+      key: SValue,
+      translateKey: (Identifier, Value[ContractId]) => Either[String, SValue],
+  )(implicit
+      ec: ExecutionContext,
+      mat: Materializer,
+  ): Future[Option[ScriptLedgerClient.ActiveContract]]
+
+  def submit(
+      actAs: OneAnd[Set, Ref.Party],
+      readAs: Set[Ref.Party],
+      commands: List[command.Command],
+      optLocation: Option[Location],
+  )(implicit
+      ec: ExecutionContext,
+      mat: Materializer,
+  ): Future[Either[StatusRuntimeException, Seq[ScriptLedgerClient.CommandResult]]]
+
+  def submitMustFail(
+      actAs: OneAnd[Set, Ref.Party],
+      readAs: Set[Ref.Party],
+      commands: List[command.Command],
+      optLocation: Option[Location],
+  )(implicit ec: ExecutionContext, mat: Materializer): Future[Either[Unit, Unit]]
+
+  def submitTree(
+      actAs: OneAnd[Set, Ref.Party],
+      readAs: Set[Ref.Party],
+      commands: List[command.Command],
+      optLocation: Option[Location],
+  )(implicit ec: ExecutionContext, mat: Materializer): Future[ScriptLedgerClient.TransactionTree]
+
+  def allocateParty(partyIdHint: String, displayName: String)(implicit
+      ec: ExecutionContext,
+      mat: Materializer,
+  ): Future[Ref.Party]
+
+  def listKnownParties()(implicit
+      ec: ExecutionContext,
+      mat: Materializer,
+  ): Future[List[PartyDetails]]
+
+  def getStaticTime()(implicit
+      ec: ExecutionContext,
+      esf: ExecutionSequencerFactory,
+      mat: Materializer,
+  ): Future[Time.Timestamp]
+
+  def setStaticTime(
+      time: Time.Timestamp
+  )(implicit ec: ExecutionContext, esf: ExecutionSequencerFactory, mat: Materializer): Future[Unit]
+
+  def tracelogIterator: Iterator[(String, Option[Location])]
+
+  def clearTracelog: Unit
+}
