+++ daml-script/test/src/com/digitalasset/daml/lf/engine/script/test/FuncWallClockIT.scala
@@ -6,7 +6,7 @@ package com.daml.lf.engine.script.test
 import java.time.Duration
 
 import com.daml.lf.data.Ref.QualifiedName
-import com.daml.lf.engine.script.ScriptTimeMode
+import com.daml.lf.engine.script.ledgerinteraction.ScriptTimeMode
 import com.daml.lf.speedy.SValue.SRecord
 
 final class FuncWallClockIT extends AbstractFuncIT {
