+++ daml-script/test/src/com/digitalasset/daml/lf/engine/script/test/TlsIT.scala
@@ -9,7 +9,7 @@ import com.daml.bazeltools.BazelRunfiles._
 import com.daml.ledger.api.testing.utils.SuiteResourceManagementAroundAll
 import com.daml.ledger.api.tls.TlsConfiguration
 import com.daml.lf.data.Ref._
-import com.daml.lf.engine.script.ScriptTimeMode
+import com.daml.lf.engine.script.ledgerinteraction.ScriptTimeMode
 import org.scalatest.matchers.should.Matchers
 import org.scalatest.wordspec.AsyncWordSpec
